Исправлена сигнатура функции

+++ tools/reability_evaluation_models/static/nelson_corcoran.c
@@ -1,8 +1,6 @@
 #include "nelson_corcoran.h"
 
-#include <stdint.h>
-
-uint8_t nelson_corcoran_reability(float* reability, uint32_t all_runs, uint32_t failed_runs) {
+float nelson_corcoran_reability(float* reability, uint32_t all_runs, uint32_t failed_runs) {
     uint8_t status = 1;
     *reability = 0.0f;
